# api/openapi.yaml
...
openapi: 3.0.0
info:
  title: MIRACUM Mapping Tool Backend
  # summary: A MIRACUM Mapping Tool Backend
  description: This is the backend for the MIRACUM Mapping Tool. It provides an API to manage projects with their corresponding mappings. Code Systems can be managed as well. The API uses Keycloak for authentication and authorization.
  # termsOfService: "https://miracum.org"
  # contact:
  #   name: API Support
  #   url: "https://miracum.org"
  #   email: support@miracum.org
  license:
    name: GPL-3.0 License
    # identifier: GPL-3.0-or-later
    url: "https://www.gnu.org/licenses/gpl-3.0.html"
  version: "0.1.0"
# servers:
#   - url: http://localhost:8080
#     description: Development server
#   - url: https://api.miracum.org
#     description: Production server
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
...
paths:
...
  /projects:
    get:
      tags:
        - project
      summary: Get all projects
      description: Get all projects with pagination. Sorting can also be specified by query parameters.
      operationId: getAllProjects
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - name: sortBy
          in: query
          description: Field to sort sortBy
          required: false
          schema:
            type: string
            default: id
            enum: [name, dateCreated, id]
        - $ref: "#/components/parameters/sortOrder"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Project"
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "401":
          "$ref": "#/components/responses/UnauthorizedError"
        "404":
          description: Page not found (e.g., page number too high)
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
    post:
      tags:
        - project
      summary: Create a new project
      description: Create a new project with permissions and code system roles (the columns in the mapping table). Code system roles can not be changed after creation.
      parameters: []
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateProjectDetails" # id prohibited/ignored, CodeSystemRoles required, Permissions optional
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetails" # id required
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "401":
          "$ref": "#/components/responses/UnauthorizedError"
        "422":
          description: Validation Exception
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
      security:
        - OAuth2: [admin]
        - BearerAuth: [admin]
    put:
      tags:
        - project
      summary: Update the project information
      description: Update the project information. Does not update the permissions for the users or the code system roles (the columns in the mapping table).
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "401":
          "$ref": "#/components/responses/UnauthorizedError"
        "403":
          "$ref": "#/components/responses/ForbiddenError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "422":
          description: Validation Exception # validate if id is provided in requestBody, it matches with the id in the path
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
  /projects/{project_id}:
    get:
      tags:
        - project
      summary: Get project details
      description: Get the project details by ID including the list of project permissions and code system roles (the columns in the mapping table)
      parameters:
        - $ref: "#/components/parameters/project_id"
      operationId: getProject
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProjectDetails" # project id optional, permissions required
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "401":
          "$ref": "#/components/responses/UnauthorizedError"
        "403":
          "$ref": "#/components/responses/ForbiddenError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
    delete:
      tags:
        - project
      summary: Delete a project
      description: Delete a project by ID. This will also delete all mappings and elements of the project as well as the permissions for the project and the code system roles.
      parameters:
        - $ref: "#/components/parameters/project_id"
      operationId: deleteProject
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Project"
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "401":
          "$ref": "#/components/responses/UnauthorizedError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
      security:
        - OAuth2: [admin]
        - BearerAuth: [admin]
...
  /codesystems/{codesystem_id}/concepts:
    get:
      tags:
        - concept
      summary: Get all concepts for a code system by ID
      description: Get all concepts for a code system by ID. Paging and sorting can be specified by query parameters.
      operationId: getAllConcepts
      parameters:
        - $ref: "#/components/parameters/codesystem_id"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - name: sortBy
          in: query
          description: Field to sort sortBy
          required: false
          schema:
            type: string
            default: code
            enum: [code, meaning]
        - $ref: "#/components/parameters/sortOrder"
        - name: codeSearch
          in: query
          description: search for the code
          required: false
          schema:
            type: string
        - name: meaningSearch
          in: query
          description: search for meaning
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Concept" # id optional
        "400":
          "$ref": "#/components/responses/BadRequestError"
        "404":
          description: CodeSystem not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          "$ref": "#/components/responses/InternalServerError"
components:
  parameters:
    project_id:
      name: project_id
      in: path
      required: true
      description: The ID of the project
      schema:
        type: integer
        format: int32
        minimum: 1
...
    codesystem_id:
      name: codesystem_id
      in: path
      required: true
      description: The ID of the Codesystem
      schema:
        type: integer
        format: int32
        minimum: 1
    page:
      name: page
      in: query
      description: Page number (must be a positive integer)
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    pageSize:
      name: pageSize
      in: query
      description: Number of items per page (minimum 1, maximum 100)
      required: false
      schema:
        type: integer
        default: 100000
        minimum: 1
    sortOrder:
      name: sortOrder
      in: query
      description: Order of sorting (asc or desc)
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  schemas:
    BaseProject:
      type: object
      required:
        - name
        - version
        - description
        - equivalence_required
        - status_required
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        equivalence_required:
          type: boolean
        status_required:
          type: boolean
    UpdateProject:
      allOf:
        - $ref: "#/components/schemas/BaseProject"
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: int32
    Project:
      allOf:
        - $ref: "#/components/schemas/UpdateProject"
        - type: object
          required:
            - modified
            - created
          properties:
            modified:
              type: string
            created:
              type: string
    CreateProjectDetails:
      allOf:
        - $ref: "#/components/schemas/BaseProject"
        - type: object
          required:
            - code_system_roles
            - project_permissions
          properties:
            project_permissions:
              type: array
              items:
                $ref: "#/components/schemas/SendProjectPermission"
            code_system_roles:
              type: array
              items:
                $ref: "#/components/schemas/CreateCodeSystemRole"
    ProjectDetails:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required:
            - code_system_roles
          properties:
            project_permissions:
              type: array
              items:
                $ref: "#/components/schemas/ProjectPermission"
            code_system_roles:
              type: array
              items:
                $ref: "#/components/schemas/CodeSystemRole"
    SendProjectPermission:
      type: object
      required:
        - user_id
        - role
      properties:
        user_id:
          type: string # uuid
        role:
          type: string
          enum: [reviewer, project_owner, editor]
    ProjectPermission:
      allOf:
      - $ref: "#/components/schemas/SendProjectPermission"
      - type: object
        required:
          - user_name
        properties:
          user_name:
            type: string
    CreateCodeSystemRole:
      type: object
      required:
       - name
       - system
       - type
      properties:
        name:
          type: string
        system:
          type: integer
          format: int32
        type:
          type: string
          enum: [source, target]
    UpdateCodeSystemRole:
        type: object
        required:
          - id
          - name
          - type
        properties:
          id:
            type: integer
            format: int32
          name:
            type: string
          type:
            type: string
            enum: [source, target]
    CodeSystemRole:
      allOf:
        - $ref: "#/components/schemas/UpdateCodeSystemRole"
        - type: object
          required:
            - system
          properties:
            system:
              type: object
              required:
                - id
                - name
                - version
              properties:
                id:
                  type: integer
                  format: int32
                name:
                  type: string
                version:
                  type: string     
...
    Concept:
      type: object
      required:
        - id
        - code
        - meaning
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        meaning:
          type: string
...
    ErrorResponse:
      type: string
  responses:
...
    BadRequestError:
      description: Invalid Request (e.g., negative id supplied, invalid body (e.g. missing field) or query parameters)
      content:
        application/json:
          schema:
            type: string
    InternalServerError:
      description: Internal Server Error due to database or implementation errors
      content:
        application/json:
          schema:
            type: string
    UnauthorizedError:
      description: Unauthorized (missing or invalid credentials)
      content:
        application/json:
          schema:
            type: string
    ForbiddenError:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            type: string
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/auth/realms/master/protocol/openid-connect/auth
          tokenUrl: http://localhost:8081/auth/realms/master/protocol/openid-connect/token
          scopes:
            admin: grants admin access
            member: grants normal access
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  # - OAuth2: [normal, admin]
  - BearerAuth: [normal, admin]